修复上下架设置邮费判断逻辑,修改店铺相关涉及价格的类型转换

See merge request health-huihui/service/health-mall/health-mall-product!21

+++ health-mall-product-model/src/main/java/com/drstrong/health/product/model/entity/store/StoreThreeRelevanceEntity.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * @author lsx
@@ -44,7 +43,7 @@ public class StoreThreeRelevanceEntity extends BaseEntity implements Serializabl
     /**
      * 三方药店商品进货价
      */
-    private BigDecimal threePurchasePrice;
+    private Integer threePurchasePrice;
 
 
     /**
